refactor: 시간 관련 유틸 통합
